Extracting delegateprovider objects to context for DI/testing
