+++ build.gradle
@@ -17,7 +17,7 @@
  */
 
 buildscript {
-  ext.kotlin_version = '1.1.61'
+  ext.kotlin_version = '1.2.0'
   ext.sourceCompatibility = 1.8
   ext.clientId = 'ktq-client'
   ext.projectDescription = 'a kotlin client for graphql'
